+++ examples/recv-iterate-ts.ts
@@ -1,4 +1,6 @@
-import {DgramAsPromised} from "../src/dgram-as-promised"
+#!/usr/bin/env -S node --experimental-specifier-resolution=node --no-warnings --loader ts-node/esm
+
+import {DgramAsPromised} from "../src/dgram-as-promised.js"
 
 const port = Number(process.argv[2]) || 0
 
@@ -9,7 +11,7 @@ async function main(): Promise<void> {
   console.info(`Socket is listening on ${address.address}:${address.port}`)
 
   for await (const packet of socket) {
-    console.info(packet)
+    console.log(packet)
     if (packet.msg.indexOf(4) !== -1) {
       await socket.close()
     }