+++ test/bind.ts
@@ -1,13 +1,13 @@
-import chai, {expect} from "chai"
+import {expect, use as chaiUse} from "chai"
 
 import chaiAsPromised from "chai-as-promised"
-chai.use(chaiAsPromised)
+chaiUse(chaiAsPromised)
 
-import dgramAsPromised, {SocketAsPromised} from "../src/dgram-as-promised"
+import dgramAsPromised, {SocketAsPromised} from "../src/dgram-as-promised.js"
 
-import {After, Feature, Given, Scenario, Then, When} from "./lib/steps"
+import {After, Feature, Given, Scenario, Then, When} from "./lib/steps.js"
 
-import mockDgram from "./lib/mock-dgram"
+import * as mockDgram from "./lib/mock-dgram.js"
 
 Feature("Test dgram-as-promised module for bind method", () => {
   Scenario("Bind socket", () => {
@@ -43,7 +43,7 @@ Feature("Test dgram-as-promised module for bind method", () => {
     After(async () => {
       try {
         await socket.close()
-      } catch (e) {
+      } catch (_e) {
         // ignore
       }
     })
@@ -70,7 +70,7 @@ Feature("Test dgram-as-promised module for bind method", () => {
     After(async () => {
       try {
         await socket.close()
-      } catch (e) {
+      } catch (_e) {
         // ignore
       }
     })