+++ test/send.ts
@@ -1,13 +1,13 @@
-import chai, {expect} from "chai"
+import {expect, use as chaiUse} from "chai"
 
 import chaiAsPromised from "chai-as-promised"
-chai.use(chaiAsPromised)
+chaiUse(chaiAsPromised)
 
-import dgramAsPromised, {SocketAsPromised} from "../src/dgram-as-promised"
+import dgramAsPromised, {SocketAsPromised} from "../src/dgram-as-promised.js"
 
-import {After, And, Feature, Given, Scenario, Then, When} from "./lib/steps"
+import {After, And, Feature, Given, Scenario, Then, When} from "./lib/steps.js"
 
-import mockDgram from "./lib/mock-dgram"
+import * as mockDgram from "./lib/mock-dgram.js"
 
 Feature("Test dgram-as-promised module for send method", () => {
   Scenario("Send datagram", () => {
@@ -42,7 +42,7 @@ Feature("Test dgram-as-promised module for send method", () => {
       try {
         await socket.close()
       } catch (e) {
-        error = e
+        error = e as Error
       }
     })
 
@@ -75,7 +75,7 @@ Feature("Test dgram-as-promised module for send method", () => {
       try {
         await socket.send("", 0, 0, 0, "")
       } catch (e) {
-        error = e
+        error = e as Error
       }
     })
 
@@ -86,7 +86,7 @@ Feature("Test dgram-as-promised module for send method", () => {
     After(async () => {
       try {
         await socket.close()
-      } catch (e) {
+      } catch (_e) {
         // ignore
       }
     })